refactor(CI): split "Build" workflow into "Go Tests" and "Acceptance"
These can be run in parallel and the name "Build" was quite misleading.
I've also changed the `concurrency` settings that we needed for
acceptance tests (to avoid running two at the same time) not to cancel
the one in progress when a new one is scheduled. The new one will wait
in line instead. This is to avoid issues with leaving the acceptance
test environment in a dirty state after a test gets cancelled in the
middle.