feat(tests): run acceptance tests in parallel and avoid cancelling them (#570)
* feat(build): test and testacc targets to depend on build

When trying to reproduce a CI failure locally we've noticed that `make
test` only failed when it was run after `make build` (because `make
build` runs code generation based on latest API definition). To avoid
accidentally not building before testing, let's make `test` depend on
`build`. Same for `testacc`.

* refactor(CI): split "Build" workflow into "Go Tests" and "Acceptance"

These can be run in parallel and the name "Build" was quite misleading.
I've also changed the `concurrency` settings that we needed for
acceptance tests (to avoid running two at the same time) not to cancel
the one in progress when a new one is scheduled. The new one will wait
in line instead. This is to avoid issues with leaving the acceptance
test environment in a dirty state after a test gets cancelled in the
middle.

* WIP: revert concurrency settings temporarily

To avoid interfering with other tests potentially in progress while
working on this PR. Will revert this before merging.

* feat: add cloud-specific prefixes to acceptance tests

And remove one that did both AKS and GKE testing but was skipped. The
idea with this change is to run the different clouds in parallel.

* feat(CI): split acceptance test job per cloud

* chore: linting fix

* fix(CI): don't cancel other jobs if one fails

* feat(ci): always run acceptance tests

We'd like to make these mandatory for merging PRs, and for that we'd
need to always run them.